backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic[email]
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.6
